Simple exclusive lock around Zendesk sync
Simple attempt at reducing the likelihood of multiple jobs being in
process at the same time for the same user.

With rate limits being observed it's more likely we'll receive multiple
webhooks for the same user over a sustained period of time. This would
result in multiple workers processing for the same user at the same
time, degrading performance for other users.

To try and reduce the likelihood of this occurring, a simple (read
flawed) exclusive lock is set by workers before performing a sync and
released afterwards.